refactor: make BaseSubModelCreateView more flexible
Allow the developer to declare the exact model fields that relate to the
parent model, along with the name it should be passed to context
(parent_object as default).

In case the developer omits the FK fields, provide some "magic" to locate
the first field name that relates the model with its parent, using the model
_meta API.

In the current state, only ForeignKey fields are supported.