chore: update sequence numbers of regions and communes after importing fixtures
Fixtures are providing their own IDs bypassing the sequence.
If a user wants to add a new region, he/she would face an
IntegrityError because nextval() is returning a used value.

At migration time, run setval() with the value of the last
inserted row so nextval() returns a valid value.