feat: add script to build an image manually
Since deployment images store commit hash and build date in their environment variables, a helper command is needed to pass the proper build arguments to Docker.

Also, exclude this helper along with quickstart.sh and reset.sh, commands that should not be present in production images in case the container is compromised.