feat: non-blocking API client methods and tasks
Use celery to run blocking requests and process callbacks.

When  using this mechanism, client configuration is dumped and sent over a message broker to the worker, where the client is instantiated again and used to make the request. A caveat of this implementation is that authentication classes cannot be serialized yet.