fix: add finalizer to client fixtures that clears RequestMiddleware locals
Thread locals are used to expose the currently logged user so audit_log can retrieve the user to create log objects.
Client fixtures not only return a Client object but also create a user and logs it in, meaning thread locals were being set without being cleaned after test ended. This caused tests not requiring those fixtures to fail as audit_log was attempting to use a user that not longer exists.